[refactor][blockaccess] Refactor s3 code.
