+++ confv2/client.conf
@@ -99,9 +99,9 @@ s3.verify_SSL=False
 s3.maxConnections=32
 s3.connectTimeout=60000
 s3.requestTimeout=10000
-s3.use_crt_client=false
+s3.use_crt_client=true
 # this only work when use_crt_client is false
-s3.use_thread_pool=true
+s3.use_thread_pool=false
 # this only work when use_crt_client is false and use_thread_pool is true
 s3.async_thread_num_in_thread_pool=256
 s3.enableTelemetry=false
