--- src/blockaccess/block_accesser_factory.cpp
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2025 dingodb.com, Inc. All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "blockaccess/block_accesser_factory.h"
-
-#include <memory>
-
-#include "blockaccess/block_accesser.h"
-
-namespace dingofs {
-namespace blockaccess {
-
-BlockAccesserUPtr BlockAccesserFactory::NewBlockAccesser(
-    const BlockAccessOptions& options) {
-  return std::make_unique<blockaccess::BlockAccesserImpl>(options);
-}
-
-BlockAccesserSPtr BlockAccesserFactory::NewShareBlockAccesser(
-    const BlockAccessOptions& options) {
-  return std::make_shared<blockaccess::BlockAccesserImpl>(options);
-}
-
-}  // namespace blockaccess
-}  // namespace dingofs
