+++ src/metaserver/compaction/s3compact_worker.h
@@ -33,7 +33,6 @@
 
 #include "absl/types/optional.h"
 #include "blockaccess/accesser_common.h"
-#include "blockaccess/block_accesser_factory.h"
 #include "metaserver/compaction/s3compact.h"
 #include "utils/interruptible_sleeper.h"
 
@@ -61,7 +60,6 @@ struct S3CompactWorkerContext {
 
 struct S3CompactWorkerOptions {
   blockaccess::BlockAccessOptions block_access_opts;
-  std::shared_ptr<blockaccess::BlockAccesserFactory> block_accesser_factory;
 
   FsInfoCache* fs_info_cache;
