+++ src/metaserver/partition_cleaner_common.h
@@ -20,7 +20,6 @@
 #include <memory>
 
 #include "blockaccess/accesser_common.h"
-#include "blockaccess/block_accesser_factory.h"
 #include "metaserver/s3/metaserver_s3_adaptor.h"
 #include "stub/rpcclient/mds_client.h"
 
@@ -31,7 +30,6 @@ struct PartitionCleanOption {
   uint32_t scanPeriodSec;
   uint32_t inodeDeletePeriodMs;
   blockaccess::BlockAccessOptions block_access_options;
-  std::shared_ptr<blockaccess::BlockAccesserFactory> block_accesser_factory;
   S3ClientAdaptorOption s3_client_adaptor_option;
   std::shared_ptr<stub::rpcclient::MdsClient> mdsClient;
 };
